implemented alarm clock functionality
Added a UI text for displaying a message on screen.

Appended the UnityEngine.UI directive in the script.

Added a [SerializeField] private Text var in the script and, after saving, dragged the text onto the new field on the script component of the circle.

Visuals: added a sprite mask to the background circle in case a larger image is used and should be clipped by the circle´s perimeter.

The "alarm" consists of a simple if-condition in the Update method, checking if datetime.seconds == value && datetime.minutes == value (as an example - hours can be additionally checked as well, of course, day month and year too).

If the condition is met, the text.gameObject is set to active with setActive(true), and displayed on screen.

This is just a test example, the if could also trigger some annoying audio cue or start your laundromat :)
go create ...

That´s pretty much it.