new alarm functionality and minor fixes
Removed smooth motion of minutes and seconds pointers in the analog clock script.

Added a new AlarmManager script containing the alarm clock functionality.

Added to the scene UI inputs for hour and minute, a  button for (re)setting the alarm, a play button for previewing the alarm sound,
sliders for pitch and volume and a text for displaying the name of the audio clip used, also a text which displays the current time in numbers.